refactor: transfer work with ls to form component
